refactor:#4 뒤로가기, 페이지네이션 컴포넌트 추가
