+++ src/app/category/layout.tsx
@@ -1,0 +1,16 @@
+import BackButton from '@/components/back-button';
+
+const CategoryLayout = ({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) => {
+  return (
+    <>
+      <BackButton />
+      {children}
+    </>
+  );
+};
+
+export default CategoryLayout;
