+++ src/app/category/popular/page.tsx
@@ -1,5 +1,5 @@
-import BackButton from '@/components/back-button';
-import { Movie } from '@/types/category/now-playing';
+import PageNationBtn from '@/components/pageNationBtn';
+import { Movie } from '@/types/category/movie';
 import { getPopular } from '@/utils/tmdb/serverApi';
 import React from 'react';
 
@@ -7,12 +7,12 @@ const popularPage = async () => {
   const { results } = await getPopular();
   return (
     <>
-      <BackButton />
       <div>
         {results.map((movie: Movie) => {
           return <div key={movie.id}>{movie.title}</div>;
         })}
       </div>
+      <PageNationBtn />
     </>
   );
 };
