+++ src/app/category/top-rated/page.tsx
@@ -1,5 +1,5 @@
-import BackButton from '@/components/back-button';
-import { Movie } from '@/types/category/now-playing';
+import PageNationBtn from '@/components/pageNationBtn';
+import { Movie } from '@/types/category/movie';
 import { getTopRated } from '@/utils/tmdb/serverApi';
 import React from 'react';
 
@@ -7,12 +7,12 @@ const topRatedPage = async () => {
   const { results } = await getTopRated();
   return (
     <>
-      <BackButton />
       <div>
         {results.map((movie: Movie) => {
           return <div key={movie.id}>{movie.title}</div>;
         })}
       </div>
+      <PageNationBtn />
     </>
   );
 };
