+++ src/app/category/upcoming/page.tsx
@@ -1,18 +1,19 @@
-import BackButton from '@/components/back-button';
-import { Movie } from '@/types/category/now-playing';
+import PageNationBtn from '@/components/pageNationBtn';
+import { Movie } from '@/types/category/movie';
 import { getUpcoming } from '@/utils/tmdb/serverApi';
 import React from 'react';
 
 const upcomingPage = async () => {
   const { results } = await getUpcoming();
+
   return (
     <>
-      <BackButton />
       <div>
         {results.map((movie: Movie) => {
           return <div key={movie.id}>{movie.title}</div>;
         })}
       </div>
+      <PageNationBtn />
     </>
   );
 };
