+++ src/components/pageNationBtn.tsx
@@ -1,0 +1,13 @@
+'use client';
+
+const PageNationBtn = () => {
+  return (
+    <div className='flex gap-8'>
+      <button>이전</button>
+      <p>현재페이지</p>
+      <button>다음</button>
+    </div>
+  );
+};
+
+export default PageNationBtn;
